@@ -1,80 +1,53 @@
 {
-  "name": "chubby",
+  "name": "admin",
+  "private": true,
   "version": "1.0.0",
   "scripts": {
-    "serve": "vue-cli-service serve --open",
-    "build": "vue-cli-service build",
-    "lint": "vue-cli-service lint --fix",
-    "dev": "npm run serve",
-    "i18n:report": "vue-cli-service i18n:report --src './src/**/*.?(js|vue)' --locales './src/locales/**/*.json'",
-    "start": "npm run serve",
-    "test:unit": "vue-cli-service test:unit"
+    "dev": "vite",
+    "build": "vue-tsc --noEmit && vite build",
+    "preview": "vite preview"
   },
   "dependencies": {
-    "@d2-projects/d2-crud": "^2.1.2",
-    "@d2-projects/vue-table-export": "^1.0.1",
-    "@d2-projects/vue-table-import": "^1.0.0",
-    "axios": "^0.19.0",
+    "@element-plus/icons-vue": "^1.1.4",
+    "axios": "^0.26.0",
     "better-scroll": "^1.15.2",
     "clipboard-polyfill": "^2.8.1",
-    "countup.js": "^2.0.4",
-    "dayjs": "^1.8.14",
-    "echarts": "^4.7.0",
-    "element-ui": "^2.13.2",
+    "core-js": "^3.8.3",
+    "dayjs": "^1.10.8",
+    "element-plus": "^2.1.2",
     "flex.css": "^1.1.7",
     "fuse.js": "^3.4.5",
-    "github-markdown-css": "^3.0.1",
-    "highlight.js": "^9.15.8",
     "hotkeys-js": "^3.6.11",
-    "js-cookie": "^2.2.0",
-    "lodash": "^4.17.11",
-    "lowdb": "^1.0.0",
-    "marked": "^0.6.3",
-    "mavon-editor": "^2.9.0",
+    "js-cookie": "^3.0.1",
+    "lodash": "^4.17.21",
+    "lowdb": "1.0.0",
+    "marked": "^4.0.12",
     "mockjs": "^1.1.0",
-    "node-sass": "^4.12.0",
+    "node-sass": "^7.0.1",
     "nprogress": "^0.2.0",
+    "path": "^0.12.7",
     "portfinder": "^1.0.21",
-    "qs": "^6.7.0",
-    "quill": "^1.3.6",
-    "screenfull": "^4.2.0",
+    "qs": "^6.10.3",
+    "sass": "^1.49.9",
+    "sass-loader": "^12.6.0",
     "simplemde": "^1.11.2",
     "socket.io-client": "^2.3.0",
     "sockjs-client": "^1.4.0",
     "stompjs": "^2.3.3",
+    "tiny-emitter": "^2.1.0",
     "ua-parser-js": "^0.7.20",
-    "v-charts": "^1.19.0",
+    "unplugin-auto-import": "^0.6.4",
+    "unplugin-element-plus": "^0.3.2",
+    "unplugin-vue-components": "^0.18.0",
     "v-contextmenu": "^2.8.0",
-    "vue": "^2.6.10",
-    "vue-grid-layout": "^2.3.4",
-    "vue-i18n": "^8.11.2",
-    "vue-json-tree-view": "^2.1.4",
-    "vue-router": "^3.1.6",
-    "vue-socket.io": "^3.0.7",
-    "vue-splitpane": "^1.0.4",
-    "vue-ueditor-wrap": "^2.4.1",
-    "vuex": "^3.1.1"
+    "vue": "^3.2.25",
+    "vue-i18n": "^9.2.0-beta.30",
+    "vue-router": "^4.0.13",
+    "vuex": "^4.0.2",
+    "webpack-theme-color-replacer": "^1.3.26"
   },
   "devDependencies": {
-    "@kazupon/vue-i18n-loader": "^0.3.0",
-    "@vue/cli-plugin-babel": "^3.8.0",
-    "@vue/cli-plugin-eslint": "^3.8.0",
-    "@vue/cli-plugin-unit-jest": "^3.8.0",
-    "@vue/cli-service": "^3.8.0",
-    "@vue/eslint-config-standard": "^4.0.0",
-    "@vue/test-utils": "1.0.0-beta.29",
-    "babel-eslint": "^10.0.2",
-    "babel-jest": "^24.8.0",
-    "compression-webpack-plugin": "^3.1.0",
-    "eslint": "^5.16.0",
-    "eslint-plugin-vue": "^5.2.2",
-    "sass": "^1.21.0",
-    "sass-loader": "^7.1.0",
-    "serialize-javascript": "^2.1.1",
-    "svg-sprite-loader": "^4.1.6",
-    "text-loader": "0.0.1",
-    "vue-cli-plugin-i18n": "^0.6.0",
-    "vue-template-compiler": "^2.6.10",
-    "webpack-theme-color-replacer": "^1.2.15"
+    "@vitejs/plugin-vue": "^2.2.4",
+    "vite": "^2.8.6"
   }
 }
