@@ -1,65 +1,55 @@
 <template>
     <div
-        class="chubby-layout-header-aside-group"
+        class="admin-layout-header-aside-group"
         :style="styleLayoutMainGroup"
-        :class="{grayMode: grayActive}"
     >
         <!-- 半透明遮罩 -->
-        <div class="chubby-layout-header-aside-mask"></div>
+        <div class="admin-layout-header-aside-mask"></div>
         <!-- 主体内容 -->
         <div
-            class="chubby-layout-header-aside-content"
+            class="admin-layout-header-aside-content"
             flex="dir:top"
         >
             <!-- 顶栏 -->
             <div
-                class="chubby-theme-header"
-                :style="{
-                    opacity: this.searchActive ? 0.5 : 1
-                }"
+                class="admin-theme-header"
+                :style="{opacity: 1}"
                 flex-box="0"
                 flex
             >
                 <div
                     class="logo-group"
-                    :style="{width: asideCollapse ? asideWidthCollapse : asideWidth}"
+                    :style="{ width: asideCollapse ? asideWidthCollapse : asideWidth }"
                     flex-box="0"
                 >
                     <img
                         v-if="asideCollapse"
-                        :src="`${$baseUrl}image/theme/${themeActiveSetting.name}/logo/icon-only.png`"
-                    >
+                        :src="`${baseUrl}image/theme/${themeActiveSetting.name}/logo/icon-only.png`"
+                    />
                     <img
                         v-else
-                        :src="`${$baseUrl}image/theme/${themeActiveSetting.name}/logo/all.png`"
-                    >
+                        :src="`${baseUrl}image/theme/${themeActiveSetting.name}/logo/all.png`"
+                    />
                 </div>
                 <div
                     class="toggle-aside-btn"
                     @click="handleToggleAside"
                     flex-box="0"
                 >
-                    <chubby-icon name="bars" />
+                    <admin-icon name="bars" />
                 </div>
-                <chubby-menu-header flex-box="1" />
+                <div flex-box="1"></div>
                 <!-- 顶栏右侧 -->
                 <div
-                    class="chubby-header-right"
+                    class="admin-header-right"
                     flex-box="0"
                 >
-                    <!-- 如果你只想在开发环境显示这个按钮请添加 v-if="$env === 'development'" -->
-                    <!-- <chubby-header-search @click="handleSearchClick"/> -->
-                    <chubby-header-fullscreen />
-                    <chubby-header-url />
-                    <chubby-header-theme />
-                    <!-- <chubby-header-locales/> -->
-                    <!-- <chubby-header-color/> -->
-                    <chubby-header-user />
+                    <admin-header-user />
                 </div>
             </div>
             <!-- 下面 主体 -->
             <div
-                class="chubby-theme-container"
+                class="admin-theme-container"
                 flex-box="1"
                 flex
             >
@@ -67,175 +57,83 @@
                 <div
                     flex-box="0"
                     ref="aside"
-                    class="chubby-theme-container-aside"
+                    class="admin-theme-container-aside"
                     :style="{
                         width: asideCollapse ? asideWidthCollapse : asideWidth,
-                        opacity: this.searchActive ? 0.5 : 1
+                        opacity: 1
                     }"
                 >
-                    <chubby-menu-side />
+                    <admin-menu-side />
                 </div>
                 <!-- 主体 -->
                 <div
-                    class="chubby-theme-container-main"
+                    class="admin-theme-container-main"
                     flex-box="1"
                     flex
                 >
-                    <!-- 搜索 -->
-                    <!-- <transition name="fade-scale">
-                <div v-if="searchActive" class="chubby-theme-container-main-layer" flex>
-                <chubby-panel-search
-                    ref="panelSearch"
-                    @close="searchPanelClose"/>
-                </div>
-            </transition> -->
                     <!-- 内容 -->
-                    <transition name="fade-scale">
+                    <!-- <transition name="fade-scale"> -->
+                    <div
+                        class="admin-theme-container-main-layer"
+                        flex="dir:top"
+                    >
+                        <!-- tab -->
                         <div
-                            v-if="!searchActive"
-                            class="chubby-theme-container-main-layer"
-                            flex="dir:top"
+                            class="admin-theme-container-main-header"
+                            flex-box="0"
                         >
-                            <!-- tab -->
-                            <div
-                                class="chubby-theme-container-main-header"
-                                flex-box="0"
-                            >
-                                <chubby-tabs />
-                            </div>
-                            <!-- 页面 -->
-                            <div
-                                class="chubby-theme-container-main-body"
-                                flex-box="1"
-                            >
+                            <admin-tabs />
+                        </div>
+                        <!-- 页面 -->
+                        <div
+                            class="admin-theme-container-main-body"
+                            flex-box="1"
+                        >
+                            <router-view v-slot="{ Component }">
                                 <transition :name="transitionActive ? 'fade-transverse' : ''">
                                     <keep-alive :include="keepAlive">
-                                        <router-view />
+                                        <component :is="Component" />
                                     </keep-alive>
                                 </transition>
-                            </div>
+                            </router-view>
                         </div>
-                    </transition>
+                    </div>
+                    <!-- </transition> -->
                 </div>
             </div>
         </div>
-        <chubby-base
-            :Visible="Visible"
-            @handleClose="handleVisible"
-        ></chubby-base>
     </div>
 </template>
 
-<script>
-import chubbyMenuSide from './components/menu-side'
-import chubbyMenuHeader from './components/menu-header'
-import chubbyTabs from './components/tabs'
-import chubbyHeaderFullscreen from './components/header-fullscreen'
-import chubbyHeaderUrl from '@/pages/base-url'
-// import chubbyHeaderLocales from './components/header-locales'
-// import chubbyHeaderSearch from './components/header-search'
-import chubbyHeaderTheme from './components/header-theme'
-import chubbyHeaderUser from './components/header-user'
-import chubbyBase from '@/pages/sys/base/index'
-// import chubbyHeaderColor from './components/header-color'
-import { mapState, mapGetters, mapActions } from 'vuex'
-import mixinSearch from './mixins/search'
-export default {
-    name: 'chubby-layout-header-aside',
-    mixins: [
-        mixinSearch
-    ],
-    components: {
-        chubbyMenuSide,
-        chubbyMenuHeader,
-        chubbyTabs,
-        chubbyHeaderFullscreen,
-        chubbyHeaderUrl,
-        // chubbyHeaderLocales,
-        // chubbyHeaderSearch,
-        chubbyHeaderTheme,
-        chubbyHeaderUser,
-        chubbyBase
-        // chubbyHeaderColor
-    },
-    data () {
-        return {
-            // [侧边栏宽度] 正常状态
-            asideWidth: '200px',
-            // [侧边栏宽度] 折叠状态
-            asideWidthCollapse: '65px',
-            Visible: false
-        }
-    },
-    // sockets: {
-    //     //查看socket是否渲染成功
-    //     connect () {
-    //         console.log("链接成功")
-    //     },
-    //     disconnect () {
-    //         console.log("断开链接")
-    //     },//检测socket断开链接 
-    //     reconnect () {
-    //         console.log("重新链接")
-    //     },
-    //     //客户端接收后台传输的socket事件
-    //     my_response (data) {
-    //         console.log('my_response', data)//接收的消息
-    //         setInterval(() => {
-    //             this.$socket.emit('heart', 'client')
-    //         }, 5000)
-    //     },
-    //     error (err) {
-    //         console.log(err)
-    //     },
-    //     heart (data) {
-    //         console.log(data)
-    //     }
-    // },
-    // created () {
-    //     window.addEventListener('beforeunload', e => {
-    //         if (this.$socket) this.$socket.close()
-    //     })
-    // },
-    computed: {
-        ...mapState('chubby', {
-            keepAlive: state => state.page.keepAlive,
-            grayActive: state => state.gray.active,
-            transitionActive: state => state.transition.active,
-            asideCollapse: state => state.menu.asideCollapse
-        }),
-        ...mapGetters('chubby', {
-            themeActiveSetting: 'theme/activeSetting'
-        }),
-        /**
-         * @description 最外层容器的背景图片样式
-         */
-        styleLayoutMainGroup () {
-            return {
-                ...this.themeActiveSetting.backgroundImage ? {
-                    backgroundImage: `url('${this.$baseUrl}${this.themeActiveSetting.backgroundImage}')`
-                } : {}
+<script setup>
+import adminMenuSide from './components/menu-side'
+import adminHeaderUser from './components/header-user/index.vue'
+import adminTabs from './components/tabs/index.vue'
+import useCurrentInstance from '@/proxy'
+import { computed, onMounted, ref } from 'vue'
+import { useStore } from "vuex"
+
+const { proxy } = useCurrentInstance()
+const baseUrl = proxy.$baseUrl
+const store = useStore()
+const asideWidth = ref('200px')
+const asideWidthCollapse = ref('65px')
+const keepAlive = computed(() => store.state.page.keepAlive)
+const transitionActive = computed(() => store.state.transition.active)
+const asideCollapse = computed(() => store.state.menu.asideCollapse)
+const themeActiveSetting = computed(() => store.getters['theme/activeSetting'])
+
+const styleLayoutMainGroup = computed(() => {
+    return {
+        ...(themeActiveSetting.backgroundImage
+            ? {
+                backgroundImage: `url('${proxy.$baseUrl}${themeActiveSetting.backgroundImage}')`,
             }
-        }
-    },
-    methods: {
-        ...mapActions('chubby/menu', [
-            'asideCollapseToggle'
-        ]),
-        /**
-         * 接收点击切换侧边栏的按钮
-         */
-        handleToggleAside () {
-            this.asideCollapseToggle()
-        },
-        handleVisible (isShow) {
-            this.Visible = isShow
-        }
+            : {})
     }
+})
+
+const handleToggleAside = () => {
+    store.dispatch('menu/asideCollapseToggle')
 }
 </script>
-
-<style lang="scss">
-// 注册主题
-@import "~@/assets/style/theme/register.scss";
-</style>
