@@ -2,13 +2,13 @@ import request from '@/plugin/axios'
 
 /**
  * 获取日志列表
- * @param {*} data 
+ * @param {*} data
  */
-export function QueryLogByParam (data) {
+export function QueryLogByParam(data) {
   return request({
     url: '/v1/Log/QueryLogByParam',
     method: 'post',
     data,
-    headers: { isCheck: true }
+    headers: { isCheck: true },
   })
 }
