@@ -10,10 +10,14 @@ const cookies = {}
  */
 cookies.set = function (name = 'default', value = '', cookieSetting = {}) {
     let currentCookieSetting = {
-        expires: 1
+        expires: 1,
     }
     Object.assign(currentCookieSetting, cookieSetting)
-    Cookies.set(`chubby-${process.env.VUE_APP_VERSION}-${name}`, value, currentCookieSetting)
+    Cookies.set(
+        `admin-${import.meta.env.VITE_APP_VERSION}-${name}`,
+        value,
+        currentCookieSetting
+    )
 }
 
 /**
@@ -21,7 +25,7 @@ cookies.set = function (name = 'default', value = '', cookieSetting = {}) {
  * @param {String} name cookie name
  */
 cookies.get = function (name = 'default') {
-    return Cookies.get(`chubby-${process.env.VUE_APP_VERSION}-${name}`)
+    return Cookies.get(`admin-${import.meta.env.VITE_APP_VERSION}-${name}`)
 }
 
 /**
@@ -36,7 +40,7 @@ cookies.getAll = function () {
  * @param {String} name cookie name
  */
 cookies.remove = function (name = 'default') {
-    return Cookies.remove(`chubby-${process.env.VUE_APP_VERSION}-${name}`)
+    return Cookies.remove(`admin-${import.meta.env.VITE_APP_VERSION}-${name}`)
 }
 
 export default cookies
