@@ -3,15 +3,15 @@ import LocalStorage from 'lowdb/adapters/LocalStorage'
 import util from '@/libs/util'
 import { cloneDeep } from 'lodash'
 
-const adapter = new LocalStorage(`chubby-${process.env.VUE_APP_VERSION}`)
+const adapter = new LocalStorage(`admin-${import.meta.env.VITE_APP_VERSION}`)
 const db = low(adapter)
 
 db
-  .defaults({
-    sys: {},
-    database: {}
-  })
-  .write()
+    .defaults({
+        sys: {},
+        database: {}
+    })
+    .write()
 
 export default db
 
@@ -25,21 +25,21 @@ export default db
  * @returns {String} 可以直接使用的路径
  */
 export function pathInit ({
-  dbName = 'database',
-  path = '',
-  user = true,
-  validator = () => true,
-  defaultValue = ''
+    dbName = 'database',
+    path = '',
+    user = true,
+    validator = () => true,
+    defaultValue = ''
 }) {
-  const uuid = util.cookies.get('uuid') || 'ghost-uuid'
-  const currentPath = `${dbName}.${user ? `user.${uuid}` : 'public'}${path ? `.${path}` : ''}`
-  const value = db.get(currentPath).value()
-  if (!(value !== undefined && validator(value))) {
-    db
-      .set(currentPath, defaultValue)
-      .write()
-  }
-  return currentPath
+    const token = util.cookies.get('token') || 'test'
+    const currentPath = `${dbName}.${user ? `user.${token}` : 'public'}${path ? `.${path}` : ''}`
+    const value = db.get(currentPath).value()
+    if (!(value !== undefined && validator(value))) {
+        db
+            .set(currentPath, defaultValue)
+            .write()
+    }
+    return currentPath
 }
 
 /**
@@ -51,16 +51,16 @@ export function pathInit ({
  * @param {Object} payload user {Boolean} 是否区分用户
  */
 export function dbSet ({
-  dbName = 'database',
-  path = '',
-  value = '',
-  user = false
+    dbName = 'database',
+    path = '',
+    value = '',
+    user = false
 }) {
-  db.set(pathInit({
-    dbName,
-    path,
-    user
-  }), value).write()
+    db.set(pathInit({
+        dbName,
+        path,
+        user
+    }), value).write()
 }
 
 /**
@@ -72,19 +72,19 @@ export function dbSet ({
  * @param {Object} payload user {Boolean} 是否区分用户
  */
 export function dbGet ({
-  dbName = 'database',
-  path = '',
-  defaultValue = '',
-  user = false
+    dbName = 'database',
+    path = '',
+    defaultValue = '',
+    user = false
 }) {
-  return new Promise(resolve => {
-    resolve(cloneDeep(db.get(pathInit({
-      dbName,
-      path,
-      user,
-      defaultValue
-    })).value()))
-  })
+    return new Promise(resolve => {
+        resolve(cloneDeep(db.get(pathInit({
+            dbName,
+            path,
+            user,
+            defaultValue
+        })).value()))
+    })
 }
 
 /**
@@ -92,15 +92,15 @@ export function dbGet ({
  * @param {Object} payload user {Boolean} 是否区分用户
  */
 export function database ({
-  dbName = 'database',
-  path = '',
-  user = false,
-  validator = () => true,
-  defaultValue = ''
+    dbName = 'database',
+    path = '',
+    user = false,
+    validator = () => true,
+    defaultValue = ''
 } = {}) {
-  return new Promise(resolve => {
-    resolve(db.get(pathInit({
-      dbName, path, user, validator, defaultValue
-    })))
-  })
+    return new Promise(resolve => {
+        resolve(db.get(pathInit({
+            dbName, path, user, validator, defaultValue
+        })))
+    })
 }
