@@ -4,15 +4,26 @@ const log = {}
  * @description 返回这个样式的颜色值
  * @param {String} type 样式名称 [ primary | success | warning | danger | text ]
  */
-function typeColor (type = 'default') {
+function typeColor(type = 'default') {
   let color = ''
   switch (type) {
-    case 'default': color = '#35495E'; break
-    case 'primary': color = '#3488ff'; break
-    case 'success': color = '#43B883'; break
-    case 'warning': color = '#e6a23c'; break
-    case 'danger': color = '#f56c6c'; break
-    default:; break
+    case 'default':
+      color = '#35495E'
+      break
+    case 'primary':
+      color = '#3488ff'
+      break
+    case 'success':
+      color = '#43B883'
+      break
+    case 'warning':
+      color = '#e6a23c'
+      break
+    case 'danger':
+      color = '#f56c6c'
+      break
+    default:
+      break
   }
   return color
 }
@@ -27,7 +38,9 @@ log.capsule = function (title, info, type = 'primary') {
   console.log(
     `%c ${title} %c ${info} %c`,
     'background:#35495E; padding: 1px; border-radius: 3px 0 0 3px; color: #fff;',
-    `background:${typeColor(type)}; padding: 1px; border-radius: 0 3px 3px 0;  color: #fff;`,
+    `background:${typeColor(
+      type
+    )}; padding: 1px; border-radius: 0 3px 3px 0;  color: #fff;`,
     'background:transparent'
   )
 }
@@ -37,8 +50,8 @@ log.capsule = function (title, info, type = 'primary') {
  */
 log.colorful = function (textArr) {
   console.log(
-    `%c${textArr.map(t => t.text || '').join('%c')}`,
-    ...textArr.map(t => `color: ${typeColor(t.type)};`)
+    `%c${textArr.map((t) => t.text || '').join('%c')}`,
+    ...textArr.map((t) => `color: ${typeColor(t.type)};`)
   )
 }
 
