@@ -3,9 +3,9 @@ export const menuAside = []
 
 // 菜单 顶栏
 export const menuHeader = [
-//   {
-//     path: '/index',
-//     title: '首页',
-//     icon: 'home'
-//   }
+  //   {
+  //     path: '/index',
+  //     title: '首页',
+  //     icon: 'home'
+  //   }
 ]
