@@ -1,10 +1,10 @@
 import util from '@/libs/util'
 
 export default {
-  install (Vue) {
-    // 快速打印 log
-    Vue.prototype.$log = {
-      ...util.log
-    }
-  }
+    install (Vue) {
+        // 快速打印 log
+        Vue.config.globalProperties.$log = {
+            ...util.log,
+        }
+    },
 }
