@@ -1,7 +1,7 @@
 import util from '@/libs/util'
 
 export default {
-  install (Vue, options) {
-    Vue.prototype.$open = util.open
-  }
+    install (Vue, options) {
+        Vue.config.globalProperties.$open = util.open
+    },
 }
