@@ -1,3 +1,3 @@
-import page from './page'
+import page from './page.vue'
 
 export default page
